Move setupConntrack from server_linux.go to conntrack.go
