+++ cmd/kube-proxy/app/server_linux.go
@@ -26,11 +26,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	goruntime "runtime"
-	"time"
-
-	"github.com/google/cadvisor/machine"
-	"github.com/google/cadvisor/utils/sysfs"
 
 	v1 "k8s.io/api/core/v1"
 	utilsysctl "k8s.io/component-helpers/node/util/sysctl"
@@ -69,8 +64,8 @@ func (o *Options) platformApplyDefaults(config *proxyconfigapi.KubeProxyConfigur
 // Proxier. It should fill in any platform-specific fields and perform other
 // platform-specific setup.
 func (s *ProxyServer) platformSetup(ctx context.Context) error {
-	ct := &realConntracker{}
-	err := s.setupConntrack(ctx, ct)
+	ct := &realConntrackConfigurer{}
+	err := setupConntrack(ctx, ct, &s.Config.Linux.Conntrack)
 	if err != nil {
 		return err
 	}
@@ -294,82 +289,6 @@ func (s *ProxyServer) createProxier(ctx context.Context, config *proxyconfigapi.
 	}
 
 	return proxier, nil
-}
-
-func (s *ProxyServer) setupConntrack(ctx context.Context, ct Conntracker) error {
-	max, err := getConntrackMax(ctx, s.Config.Linux.Conntrack)
-	if err != nil {
-		return err
-	}
-	if max > 0 {
-		err := ct.SetMax(ctx, max)
-		if err != nil {
-			return err
-		}
-	}
-
-	if s.Config.Linux.Conntrack.TCPEstablishedTimeout != nil && s.Config.Linux.Conntrack.TCPEstablishedTimeout.Duration > 0 {
-		timeout := int(s.Config.Linux.Conntrack.TCPEstablishedTimeout.Duration / time.Second)
-		if err := ct.SetTCPEstablishedTimeout(ctx, timeout); err != nil {
-			return err
-		}
-	}
-
-	if s.Config.Linux.Conntrack.TCPCloseWaitTimeout != nil && s.Config.Linux.Conntrack.TCPCloseWaitTimeout.Duration > 0 {
-		timeout := int(s.Config.Linux.Conntrack.TCPCloseWaitTimeout.Duration / time.Second)
-		if err := ct.SetTCPCloseWaitTimeout(ctx, timeout); err != nil {
-			return err
-		}
-	}
-
-	if s.Config.Linux.Conntrack.TCPBeLiberal {
-		if err := ct.SetTCPBeLiberal(ctx, 1); err != nil {
-			return err
-		}
-	}
-
-	if s.Config.Linux.Conntrack.UDPTimeout.Duration > 0 {
-		timeout := int(s.Config.Linux.Conntrack.UDPTimeout.Duration / time.Second)
-		if err := ct.SetUDPTimeout(ctx, timeout); err != nil {
-			return err
-		}
-	}
-
-	if s.Config.Linux.Conntrack.UDPStreamTimeout.Duration > 0 {
-		timeout := int(s.Config.Linux.Conntrack.UDPStreamTimeout.Duration / time.Second)
-		if err := ct.SetUDPStreamTimeout(ctx, timeout); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func getConntrackMax(ctx context.Context, config proxyconfigapi.KubeProxyConntrackConfiguration) (int, error) {
-	logger := klog.FromContext(ctx)
-	if config.MaxPerCore != nil && *config.MaxPerCore > 0 {
-		floor := 0
-		if config.Min != nil {
-			floor = int(*config.Min)
-		}
-		scaled := int(*config.MaxPerCore) * detectNumCPU()
-		if scaled > floor {
-			logger.V(3).Info("GetConntrackMax: using scaled conntrack-max-per-core")
-			return scaled, nil
-		}
-		logger.V(3).Info("GetConntrackMax: using conntrack-min")
-		return floor, nil
-	}
-	return 0, nil
-}
-
-func detectNumCPU() int {
-	// try get numCPU from /sys firstly due to a known issue (https://github.com/kubernetes/kubernetes/issues/99225)
-	_, numCPU, err := machine.GetTopology(sysfs.NewRealSysFs())
-	if err != nil || numCPU < 1 {
-		return goruntime.NumCPU()
-	}
-	return numCPU
 }
 
 func getLocalDetectors(logger klog.Logger, primaryIPFamily v1.IPFamily, config *proxyconfigapi.KubeProxyConfiguration, nodePodCIDRs []string) map[v1.IPFamily]proxyutil.LocalTrafficDetector {
